requirements.txt
backports.functools-lru-cache
boto3
botocore
click
configparser
docutils
enum34
Flask
Flask-Login
Flask-WTF
gevent
greenlet
gunicorn
isort
itsdangerous
Jinja2
jmespath
lazy-object-proxy
MarkupSafe
mccabe
numpy
psycopg2-binary
PyJWT
pylint
python-dateutil
s3transfer
singledispatch
six
transforms3d
Werkzeug
wrapt
WTForms
